Update infrastructure to use SQL Server instead of SQLite.

+++ backend/src/MovieStore.Domain/Movies/Movie.cs
@@ -1,11 +1,11 @@
 namespace MovieStore.Domain.Movies;
 
-public class Movie(string title, string description, double price)
+public class Movie(string title, string description, decimal price)
 {
     public int Id { get; set; }
     public string Title { get; set; } = title;
     public string Description { get; set; } = description;
-    public double Price { get; set; } = price;
+    public decimal Price { get; set; } = price;
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     
     public List<MovieGenre> MovieGenres { get; set; }
